UI Updates made to project
Hello, I have made some changes to the battle scene GUI along with starting work on a potential vfx we can play upon the player attacking
the changes are listed below:

sprites have been created for the battle GUI:
Sprites for panels and buttons have been created and animated for the GUI

button animation and colour switching upon selection of the buttons:
I have created a simple animation controller for the gui buttons, this uses button enter and exit events to change the sprite that is being used. Hovering over a button will change the sprite so that it is more clear which button is currently being selected, in addition i have also added colour changes.

fire animation; I have created and added a fire animation that is activated and deactivated with enter and exit events. The fire animation is displayed next to the selected button

enemy HP bars:
enemy hp bars will spawn above the heads of enemies and be removed upon enemy deaths. The hp bar also updates and changes as the enemies lose health (Im currently using unitys bult in UI sprites for this, i can draw my own soon)

reshuffling of the battle scene GUI:
with the new sprites i have slightly reshuffled and resized parts of the battle scene gui so it remains all readable

simple vfx:
i have made the skeleton of a vfx that can be used when an attack is played